Initial HTML elements for my recipes webpage
The files included in this commit collectively define the layout of my
recipes website. The "index.html" files contains links to three
separate HTML documents, each of which are recipes for a particular
meal. Each recipe page consists of a header, image, description,
an unordered list of ingredients and an ordered list with the steps to
prepare the meal.